[script.program.driverselect] v0.0.6 - changes in listitems

+++ default.py
@@ -1,4 +1,5 @@
 from resources.lib.tools import *
+import os
 
 ADDON_TYPE = 'xbmc.service'
 SIGNATURE = 'driver.dvb.'
@@ -47,41 +48,44 @@ if response is not None:
         driver_module = dialogSelect(LS(30011), gui_list, preselect=preselection, useDetails=True)
         if driver_module > -1:
             writeLog('selected item: %s' % gui_list[driver_module].getProperty('addonid'), xbmc.LOGNOTICE)
+            if driver_module != preselection:
 
-            # disable old driver module(s)
+                # disable old driver module(s)
 
-            for dvbmodule in modules:
-                query = {"method": "Addons.SetAddonEnabled",
-                         "params":{"addonid": dvbmodule, "enabled": False}}
-                response = jsonrpc(query)
-                if response == 'OK':
-                    writeLog('driver module \'%s\' disabled' % (dvbmodule), xbmc.LOGNOTICE)
-                else:
-                    writeLog('could not disable driver module \'%s\'' % (dvbmodule), xbmc.LOGFATAL)
-
-            # enable new driver module
-
-            query = {"method": "Addons.SetAddonEnabled",
-                     "params": {"addonid": gui_list[driver_module].getProperty('addonid'), "enabled": True}}
-            response = jsonrpc(query)
-            if response == 'OK':
-                writeLog('driver module \'%s\' enabled' % (gui_list[driver_module].getProperty('name')), xbmc.LOGNOTICE)
-
-                # ask for reboot
-
-                yesno = dialogYesNo(LS(30012), LS(30013))
-                if yesno:
-                    query = {"method": "System.Reboot",
-                             "params":{}}
+                for dvbmodule in modules:
+                    query = {"method": "Addons.SetAddonEnabled",
+                             "params":{"addonid": dvbmodule, "enabled": False}}
                     response = jsonrpc(query)
                     if response == 'OK':
-                        writeLog('system will now reboot', xbmc.LOGNOTICE)
+                        writeLog('driver module \'%s\' disabled' % (dvbmodule), xbmc.LOGNOTICE)
                     else:
-                        writeLog('could not reboot', xbmc.LOGFATAL)
+                        writeLog('could not disable driver module \'%s\'' % (dvbmodule), xbmc.LOGFATAL)
+
+                # enable new driver module
+
+                query = {"method": "Addons.SetAddonEnabled",
+                         "params": {"addonid": gui_list[driver_module].getProperty('addonid'), "enabled": True}}
+                response = jsonrpc(query)
+                if response == 'OK':
+                    writeLog('driver module \'%s\' enabled' % (gui_list[driver_module].getProperty('name')), xbmc.LOGNOTICE)
+
+                    # ask for reboot
+
+                    yesno = dialogYesNo(LS(30012), LS(30013))
+                    if yesno:
+                        query = {"method": "System.Reboot",
+                                 "params":{}}
+                        response = jsonrpc(query)
+                        if response == 'OK':
+                            writeLog('system will now reboot', xbmc.LOGNOTICE)
+                        else:
+                            writeLog('could not reboot', xbmc.LOGFATAL)
+                    else:
+                        notify(LS(30010), LS(30014), icon=xbmcgui.NOTIFICATION_WARNING)
                 else:
-                    notify(LS(30010), LS(30014), icon=xbmcgui.NOTIFICATION_WARNING)
+                    writeLog('could not enable driver module \'%s\'' % (gui_list[driver_module].getProperty('name')), xbmc.LOGFATAL)
             else:
-                writeLog('could not enable driver module \'%s\'' % (gui_list[driver_module].getProperty('name')), xbmc.LOGFATAL)
+                writelog('module does\'nt changed, no further actions required')
         else:
             writeLog('selection aborted')
     else:
